requirements.txt
pyqt5
websocket-client
